Separate hook types into better-named files
We want all react-redux types in their own file, separate from the React ones.

Also, since these are type definitions, they should go in the type folder.
